Removed the ability to delete user with admin role, improved login layout

+++ app/Http/Controllers/UserController.php
@@ -86,6 +86,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_if($user->isAdmin(), 500, 'Cannot delete a user with role "admin"');
         $user->delete();
         return redirect(route('users.index'));
     }
